Move Octopus.Shared into Octopus.Tentacle

+++ source/Octopus.Tentacle/Commands/DeleteInstanceCommand.cs
@@ -1,7 +1,7 @@
 using System;
 using Octopus.Diagnostics;
-using Octopus.Shared.Configuration.Instances;
-using Octopus.Shared.Startup;
+using Octopus.Tentacle.Configuration.Instances;
+using Octopus.Tentacle.Startup;
 
 namespace Octopus.Tentacle.Commands
 {
@@ -19,7 +19,7 @@ namespace Octopus.Tentacle.Commands
 
         protected override void Start()
         {
-            instanceManager.DeleteInstance(instanceSelector.Current.InstanceName);
+            if (instanceSelector.Current.InstanceName != null) instanceManager.DeleteInstance(instanceSelector.Current.InstanceName);
         }
     }
 }
